requirements.txt
fastapi[all]
uvicorn
supabase
asyncpg
databases
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
PyJWT
python-dotenv
pydantic-settings
httpx
email-validator